+++ common/build.gradle.kts
@@ -1,4 +1,5 @@
 plugins {
+    id("geckolib-convention")
     id("org.spongepowered.gradle.vanilla") version "0.2.1-SNAPSHOT"
 }
