+++ fabric/build.gradle.kts
@@ -2,6 +2,7 @@ import net.fabricmc.loom.task.RemapJarTask
 import net.darkhax.curseforgegradle.TaskPublishCurseForge
 
 plugins {
+    id("geckolib-convention")
     id("com.modrinth.minotaur")
     id("net.darkhax.curseforgegradle")
     id("fabric-loom") version "1.5-SNAPSHOT"
