+++ neoforge/build.gradle.kts
@@ -1,6 +1,7 @@
 import net.darkhax.curseforgegradle.TaskPublishCurseForge
 
 plugins {
+    id("geckolib-convention")
     id("com.modrinth.minotaur")
     id("net.darkhax.curseforgegradle")
     id("net.neoforged.gradle.userdev") version "7.+"
